--- repos/paella-cookieconsent-plugin/src/i18n/en-US.json
@@ -1,3 +1,0 @@
-{
-    
-}
