--- repos/paella-cookieconsent-plugin/src/i18n/es-ES.json
@@ -1,4 +1,0 @@
-{
-    "Cookie Consent Plugin": "Consentimiento de Cookies",
-    "This plugin manages cookie consent preferences for the player, allowing users to accept or reject cookies and configure their preferences.": "Este plugin gestiona las preferencias de consentimiento de cookies para el reproductor, permitiendo a los usuarios aceptar o rechazar cookies y configurar sus preferencias."
-}
