+++ repos/paella-extra-plugins/README.md
@@ -1,29 +1,40 @@
-# paella-cookieconsent-plugin
+# paella-extra-plugins
+
+A extra plugin set for Paella Player
+
+
+## Installation
+
+Install the plugin via npm:
+
+```bash
+npm install @asicupv/paella-extra-plugins
+```
+
+
+
+# es.upv.paella.cookieConsent Plugin
 
 A plugin to manage cookie consent in Paella Player.
 
 This plugin is useful when the site integrating Paella Player does not already have a way to manage cookie consent. If the site already handles cookie consent, you should implement the `getCookieConsentFunction` from `PaellaInitParams`. For more details on how to implement this function, refer to the Paella documentation.
 
-## Usage
+## Description
 
-**Step 1:** Import the plugin context and add it to the Paella Player initialization parameters:
+The `es.upv.paella.cookieConsentPlugin` displays a cookie consent banner in the video player. 
 
-```javascript
-...
-import getCookieConsentPluginContext from 'paella-cookieconsent-plugin';
+### Step 1: Import the plugin and CSS
 
-let paella = new Paella('player-container', {
-    customPluginContext: [
-        getCookieConsentPluginContext()
-    ]
-});
-...
+```typescript
+import { CookieConsentPlugin } from '@asicupv/paella-extra-plugins';
+import '@asicupv/paella-extra-plugins/paella-extra-plugins.css';
 ```
 
-**Step 2:** Configure the `cookieConsent` property and add the plugin in the `plugins` section of the Paella Player configuration file.
 
-```json
-{
+### Step 2: Add the plugin and configure the `cookieConsent` property into the Paella Player configuration
+
+```typescript
+let paella = new Paella('player-container', {
     "cookieConsent": [
         {
             "type": "necessary",
@@ -51,20 +62,25 @@ let paella = new Paella('player-container', {
         }
     ],
     ...
-    "plugins": {
-        "es.upv.paella.cookieConsentPlugin": {
-            "enabled": true
+    plugins: [
+        {
+            plugin: CookieConsentPlugin,
+            config: {
+                enabled: true
+            }
         }
-    }
-}
+    ]
+});
 ```
 
-**Step 3:** Add the `getCookieConsentFunction` function to the `InitParams` when creating the player.
+
+
+## Step 3: Add the `getCookieConsentFunction` function to the `InitParams` when creating the player.
 
 The `getCookieConsentFunction` is used to retrieve the user's cookie consent preferences. This function must be passed as part of the initialization parameters when creating the Paella Player instance.
 
 ```javascript
-import { getCookieConsentFunction } from 'paella-cookieconsent-plugin';
+import { getCookieConsentFunction } from 'paella-extra-plugins';
 
 const player = new Paella('playerContainer', {
     getCookieConsentFunction: getCookieConsentFunction,
@@ -74,9 +90,6 @@ const player = new Paella('playerContainer', {
 
 This function ensures that the plugins and features in Paella Player respect the user's cookie preferences, such as enabling or disabling specific functionalities based on the consent type.
 
-## Description
-
-The `paella-cookieconsent-plugin` displays a cookie consent banner in the video player. 
 
 ### Icon Customization
 
@@ -121,3 +134,73 @@ The `paella-cookieconsent-plugin` displays a cookie consent banner in the video 
     }
 }
 ```
+
+### Screenshot
+![cookieconsent_view.png](cookieconsent_view.png)
+
+
+# es.upv.paella.onboarding Plugin
+
+## Description
+
+The `es.upv.paella.onboarding` provides an interactive tutorial for users to familiarize themselves with the features of the Paella Player. It uses the `shepherd.js` library to create guided tours.
+
+## Features
+
+- Step-by-step tutorial for navigating the player.
+- Highlights key elements of the player interface.
+- Option to disable the tutorial permanently.
+
+
+## Usage
+
+### Step 1: Import the plugin and CSS
+
+```typescript
+import { OnboardingPlugin } from '@asicupv/paella-extra-plugins';
+import '@asicupv/paella-extra-plugins/paella-extra-plugins.css';
+```
+
+### Step 2: Add the plugin to the Paella Player configuration
+
+```typescript
+let paella = new Paella('player-container', {
+    plugins: [
+        {
+            plugin: OnboardingPlugin,
+            config: {
+                enabled: true
+            }
+        }
+    ]
+});
+```
+
+
+## Configuration
+
+The plugin can be configured in the `config.json` file:
+
+```json
+{
+    "plugins": {
+        "es.upv.paella.onboarding": {
+            "enabled": true
+        }
+    }
+}
+```
+
+### Options
+
+- `enabled`: Enable or disable the plugin.
+
+## Screenshots
+
+### Example of a tutorial step
+![tutorial_step.png](tutorial_step.png)
+
+## Dependencies
+
+- [shepherd.js](https://shepherdjs.dev/): Library for creating guided tours.
+
