+++ repos/paella-extra-plugins/changelog.md
@@ -1,0 +1,5 @@
+# Changelog
+
+- 2.5.0:
+    * new paella-extra-plugins
+    * paella-onboarding-plugin and paella-cookieconsient-plugin moved to paella-extras-plugin
