+++ repos/paella-extra-plugins/src/i18n/all.ts
