+++ repos/paella-extra-plugins/src/i18n/es-ES.json
@@ -13,5 +13,7 @@
     "You're all set!": "¡Has terminado!",
     "You've reached the end of the tour.<br/><br/>Now you're ready to explore and use the Paella Player on your own.": "Has llegado al final del tutorial.<br/><br/>Ahora estás listo para explorar y usar el reproductor Paella por tu cuenta.",
     "This button opens a menu with multiple additional tools designed to enhance your experience.": "Este botón abre un menú con múltiples herramientas adicionales diseñadas para mejorar tu experiencia.",
-    "Here you’ll find all the available tools grouped in one place. Take a moment to explore what's available.": "Aquí encontrarás todas las herramientas disponibles agrupadas en un solo lugar. Tómate un momento para explorar lo que hay disponible."
+    "Here you’ll find all the available tools grouped in one place. Take a moment to explore what's available.": "Aquí encontrarás todas las herramientas disponibles agrupadas en un solo lugar. Tómate un momento para explorar lo que hay disponible.",
+    "Cookie Consent Plugin": "Consentimiento de Cookies",
+    "This plugin manages cookie consent preferences for the player, allowing users to accept or reject cookies and configure their preferences.": "Este plugin gestiona las preferencias de consentimiento de cookies para el reproductor, permitiendo a los usuarios aceptar o rechazar cookies y configurar sus preferencias."
 }
