+++ repos/paella-extra-plugins/src/plugins/PackagePluginModule.ts
