+++ repos/paella-extra-plugins/src/vite-env.d.ts
