--- repos/paella-onboarding-plugin/changelog.md
@@ -1,3 +1,0 @@
-# Changelog
-
-- 2.2.12: Group help is now displayed
