--- repos/paella-onboarding-plugin/src/index.ts
@@ -1,13 +1,0 @@
-import OnBoarding from './plugins/es.upv.paella.onboarding.js';
-
-export const onboardingPlugins = [
-    {
-        plugin: OnBoarding,
-        config: {
-            enabled: true
-        },
-    }
-];
-
-
-export const OnBoardingPlugin = OnBoarding;
