+++ samples/development-player-ts/package.json
@@ -17,7 +17,7 @@
     "@asicupv/paella-basic-plugins": "^2.3.0",
     "@asicupv/paella-cookieconsent-plugin": "^2.0.0",
     "@asicupv/paella-core": "^2.3.0",
-    "@asicupv/paella-onboarding-plugin": "^2.2.11",
+    "@asicupv/paella-extra-plugins": "^2.2.11",
     "@asicupv/paella-slide-plugins": "^2.3.0",
     "@asicupv/paella-user-tracking": "^2.2.2",
     "@asicupv/paella-video-plugins": "^2.2.2",
